feat: research

+++ internal/service/user.go
@@ -24,10 +24,13 @@ func (s *Service) UserList(users *[]model.User, page int, size int, total *int64
 	return nil
 }
 
-func (s *Service) UserRetrieve(user *model.User, id int64) error {
+func (s *Service) UserRetrieve(user *model.User, id uint64) error {
 	if err := s.db.Model(&model.User{}).
 		Preload("Roles").
 		First(&user, id).Error; err != nil {
+		// if err == gorm.ErrRecordNotFound {
+		// 	return nil
+		// }
 		return err
 	}
 	return nil
@@ -47,7 +50,7 @@ func (s *Service) UserUpdate(user *model.User, payload map[string]interface{}) e
 	return nil
 }
 
-func (s *Service) UserDelete(id int) error {
+func (s *Service) UserDelete(id uint64) error {
 	if err := s.db.Delete(&model.User{}, id).Error; err != nil {
 		return err
 	}
